+++ src/modifiers/blocks/Clear.spec.ts
